Transform to working as an entry point plugin
xref: xarray-contrib/xpublish#140
