+++ .pre-commit-config.yaml
@@ -25,7 +25,7 @@ repos:
     hooks:
       - id: blackdoc
 
-  - repo: https://gitlab.com/pycqa/flake8
+  - repo: https://github.com/pycqa/flake8
     rev: 3.9.0
     hooks:
       - id: flake8
