+++ xpublish_edr/__init__.py
@@ -2,9 +2,9 @@
 xpublish_edr is not a real package, just a set of best practices examples.
 """
 
-from xpublish_edr.cf_edr_router import cf_edr_router
+from xpublish_edr.plugin import CfEdrPlugin
 
-__all__ = ["cf_edr_router"]
+__all__ = ["CfEdrPlugin"]
 
 try:
     from ._version import __version__
